bottom sheet implemented

+++ app/src/main/java/com/example/milkflow/ui/SupplierFragment.kt
@@ -1,10 +1,14 @@
 package com.example.milkflow.ui
 
+import android.app.AlertDialog
+import android.app.Dialog
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.replace
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.RecyclerView
+import com.example.milkflow.AddPersonDataFragment
 import com.example.milkflow.R
 import com.example.milkflow.adapter.MilkPersonAdapter
 import com.example.milkflow.database.PersonDatabase
@@ -13,6 +17,7 @@ import com.example.milkflow.repository.MilkRepository
 import com.example.milkflow.utils.DialogUtils
 import com.example.milkflow.viewmodel.MilkViewModel
 import com.example.milkflow.viewmodel.MilkViewModelFactory
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.snackbar.Snackbar
 
 
@@ -37,7 +42,12 @@ class SupplierFragment : Fragment(R.layout.fragment_supplier) {
         binding.lifecycleOwner = viewLifecycleOwner
 
         binding.addButton.setOnClickListener {
-            DialogUtils.addPersonDialog(requireContext(), viewModel, "Supplier")
+
+
+            val supportFragmentManager = (activity as MainActivity).supportFragmentManager
+            val fragment = AddPersonDataFragment.newInstance("Supplier", "Person")
+            fragment.show(supportFragmentManager, AddPersonDataFragment::class.java.simpleName)
+
         }
 
         viewModel.getSuppliers().observe(viewLifecycleOwner) {
@@ -49,12 +59,18 @@ class SupplierFragment : Fragment(R.layout.fragment_supplier) {
         adapter = MilkPersonAdapter(
             isSupplier = true,
             onDeletePerson = { person ->
-                viewModel.delete(person)
-//            myToast(requireContext(),"${person.personName} is deleted", R.drawable.baseline_delete_24)
 
+                val deleteConfirm = AlertDialog.Builder(requireContext())
+                deleteConfirm.setTitle("Delete Record")
+                    .setMessage("Are you sure?")
+                    .setPositiveButton("Yes"){_, _ ->
+                        viewModel.delete(person)
                     Snackbar.make(view, "${person.personName} is deleted",Snackbar.LENGTH_LONG)
                         .setAction("Undo") { viewModel.insert(person) }
                         .show()
+                    }
+                    .setNegativeButton("Cancel", null)
+                    .show()
 
             }, onEditPerson = { person ->
                 DialogUtils.editPersonDialog(requireContext(), viewModel, person, "Supplier")
